requirements.txt
anthropic
area
beautifulsoup4
boto3
fastapi
geopandas
mysql-connector-python
mysql_connector_repackaged
matplotlib
numpy
pandas
pyarrow
pydantic
python-dotenv
pymysql
python-multipart
requests
selenium
shapely
sqlalchemy
tqdm
openpyxl
unidecode
uvicorn
aiohttp
nest-asyncio
pypdf
webdriver-manager